requirements.txt
arrow==0.12.1
pyTelegramBotAPI==3.6.6
PyYAML==3.13
requests==2.21.0
schedule==0.5.0
telebot==0.0.3